requirements.txt
python-dotenv
astral
fastapi
uvicorn[standard]
requests
requests-cache
retry-requests
openmeteo-requests
meteostat==1.7.6
pandas
numpy
scikit-learn
tensorflow
cryptography
kalshi_python_sync
psycopg2-binary